pyproject.toml
[project]
name = "findable-api"
version = "0.0.1"
description = "Findable Album"
authors = [
    {name = "jwkwon0817",email = "jeewon.kwon.0817@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "openai (>=1.90.0,<2.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "tortoise-orm[psycopg] (>=0.25.1,<0.26.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "tomlkit (>=0.13.3,<0.14.0)"
]

[tool.taskipy.tasks]
start = "python manage.py runserver 0.0.0.0:8088"
migrate = "python manage.py makemigrations && python manage.py migrate"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"] 

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true 

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
ruff = "^0.12.0"
aerich = "^0.9.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "app.core.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
